requirements.txt
openai>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
gspread>=6.0.0,<7.0.0
oauth2client>=4.1.3,<5.0.0
python-dotenv>=1.0.0,<2.0.0
schedule>=1.2.0,<2.0.0
elevenlabs>=2.3.0,<3.0.0
google-api-python-client>=2.172.0,<3.0.0
google-auth>=2.40.3,<3.0.0
google-auth-oauthlib>=1.2.2,<2.0.0
httplib2>=0.22.0,<1.0.0